refactor(tests): moved tests into __tests__
